+++ domain/src/main/java/com/omegar/domain/entity/CustomFieldsValue.kt
@@ -5,5 +5,5 @@ import java.io.Serializable
 interface CustomFieldsValue : Serializable {
     val name: String?
     val presentation: String?
-    val minutes: Int?
+    val minutes: Long?
 }
